+++ CHANGELOG.md
@@ -5,6 +5,7 @@ This document shows the changes per release
 ## v1.1.5 - Current Dev Branch
 
   * Standardised module configuration handling (which helps reduce the number of exceptions related to configuration)
+  * Added configuration to enable/disable and configure HTTPControl module
 
 ## v1.1.4 - 2019-10-22